internal/cli/internal/cmd: renamed to internal/cli/internal/command
