+++ internal/cli/internal/command/context.go
@@ -1,4 +1,4 @@
-package cmd
+package command
 
 import (
 	"context"
