+++ internal/cli/internal/command/context_test.go
@@ -1,4 +1,4 @@
-package cmd
+package command
 
 import (
 	"context"
